changing some definitions to test repl
