--- tests/mocha.test.ts
@@ -1,5 +1,0 @@
-// suppress console logging so output of mocha is clear
-before(function () {
-  console.log = function () {};
-});
-
